Move SubTaskfile files to the addon folder
