+++ src/components/Generator/GeneredTaskfile/addons/subTaskfile/sub-taskfile-base.sh
@@ -5,16 +5,16 @@
 
 function task:example { ## This is an example sub task
     title "Run example sub task"
+	echo -e "./SubTaskfile location: ${YELLOW}$SUBTASKFILE_DIR${RESET}"
 	echo -e "${GREEN}Success!${RESET}"
 }
 
 # =========================================================
 # SubTaskfile helper
-#
-# Note: Optional, but without this you can only run
-# `./Taskfile foo` to get help output
+# Note: Optional, if you want to have an explicit
+#       SubTaskfile help command (can't be named `help`)
 # =========================================================
 
-function task:subtask-help { ## Show all available tasks
+function task:subtask-help { ## Show all available sub-tasks
 	task:help
 }
